pyproject.toml
[tool.poetry]
name = "botw-bot"
version = "0.1.0"
description = ""
authors = ["Noah <noah@dreckbu.de>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
alembic = "^1.6.5"
asyncpg = "^0.23.0"
click = "^8.0.1"
dateparser = "^1.0.0"
jishaku = "^2.5.1"
pendulum = "^2.1.2"
Pillow = "^9.0.0"
psycopg2-binary = "^2.9.1"
PyYAML = "^5.4.1"
seaborn = "^0.11.1"
aioscheduler = {git = "https://github.com/noahkw/aioscheduler"}
gfypy = "^1.2.4"
peony-twitter = "^2.0.0"
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus"}
SQLAlchemy = "^1.4.22"
pandas = "^1.3.1"
scipy = "^1.7.3"
discord-py = "^2.1.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.13.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
